[#2696] Show draft updates in Need reporting filter
The Need reporting filter includes periods with only new and draft status updates.

Refactor hideMe to take this into account and show those updates.

This commit also includes code in UpdateForm to re-render the page after the saving of an update.
However this breaks when deleting updates, and needs a re-think where the updates state changes drive the re-render.
See the TODO comment in successCallback()