[#2684] Keep the set page size value even after filters applied
Page size is set using a `limit` query parameter. When filters are applied, a
form submission happens, and only the form data is submitted to the server. The
`limit` query parameter gets lost.  This commit adds a hidden field for the
limit parameter to the form, so that this data is retained even when form
submissions happen.

Closes #2684