[#2715] Refactor update form visibility
There can now ony be one update form visible, in the right hand pane.

This is controlled by store.ui.updateFormDisplay, the value of which is either false, when the form is closed, or the ID of the update the form displays.

Also the rendered form in the accordion is kept closed when clicking Edit update.