[#2715] Revise filtering rules
Change the rules that determine which periods are visible in each filter.

A period is now included if any update for the period falls into the filter category.

This leads to periods with updates that have different states may be visible in more than one filter view.

An update that is shown in a certain filter view but that does not "belong" to that filter, e.g. a draft update shown in the Pending approval filter, is shown in a dimmed state, and cannot be edited, regardless of the user's permissions.