[#2785] Object.assign is not implemented in IE
This commit adds a polyfill version for it in the advanced filters code to get
it working in IE. Object.assign seems to be used in the new results framework
too, but the bundles seem to have polyfill implementations in them.

Also, add a polyfill version of Promise (used by fetch.js) for browsers that
don't implement it.

Closes #2785