[#2771] Fix display of organisation names in project main partners tab
When importing organisations from IATI, the Organisation name field may have been truncated to 25 or (more seldom) 40 characters.
Therefore we check if the name is exactly 25 or 40 chars long AND is identical to the substring of long_name to that length.
If this is the case we use the long_name as there is a very high probability that the name field has been truncated.