[#2674] Allow nested posting with disaggregations
In the end-point for posting indicator_period_data_framework, allow posting
updates along with corresponding disaggregations. The nested posting is not
supported by default, by the version of DRF being used currently, and this
needed a slight hack in the data validation step to get working, correctly. Not
sure, if there's a way to do this, without the hack.q