[#2849] Fix missing timestamp and user org
When generating an update thathas no user associated with it we have no timestamp from the LogEntry and should not
try to set a timestamp to the new update.

When the RSR system user is used there is no  organisation for that user to show in the TimestampInfo component.
It's somewhat moot at this point since the organisation is hidden in the component, but since the design isn't completely done we're keeping the element and thus have to fix it when there is no org.