[#2675] Increase height of accordion scroll area
On small screens the viewport height of the scrollable accordion becomes very small.
This is because of the large vertical space taken up by the page headers an footers.

Mitigate this problem by increasing the height of the accordion viewport. This results in the page having two scrollbars, one for the accordion and one for the page. However currently this is better than almost seeing nothing at all of the accordion on small screens.

Also hide the project footer when the results tab is chosen on the project page.