[#2883] Fix hiding of indicator labels in project editor
Indicator labels should only be visible for qualitative indicators.

Check that there are any labels to choose from, i.e. that at least one of the project's organisations has labels defined. If not labels are hidden.

Also make sure that if an indicator changes type from qualitative, all labels for that indicator are deleted.