[#2859 #2903] Fix image removal in indicator update form
This is a functional fix. By setting the z-index on the div that displays the "remove image" text it comes on top of the input element allowing it to be shown on hover and clicked on.

However the text itself doesn't align vertically corract, leading to it being below the image if the image height is small.