[#2919] Fix updates not updating in the accordion
When saving indicator updates in some cases they are not correctly updated in the accordion UI.

This is due to a bug in the App.onClose() method that updates the store with the wrong data.

Fix by only updating the UI when needed, i.e. when the user closes the update form via the Close button (the X). When the update is saved the store is updated by the as part of the saving of th update to the backend and does not need the reverting that happens when closing the update.