[#2928] Update iati_codelist_generator script
Refactor the script to be able to handle the adding of calls to the translation machinery for select fields of select codelists. This is based on the translated_codelists dict that defines which codelists and fields should be tagged for translation.

Add handling of codelist XML with tranlsations. They are simply ignored for now, since they seem to be too few to be of use.

The script is also refactored some to (hopefully) be easier to extend with other customization based on codelists and/or fields.