[#2946] Remove calls to parent.save()
Calling child.save and parent.save in a transaction is causing problems because
the parent.save method in-turn calls `update_or_create`, and since we are just
adding the reference to the parent in the current transaction, the child is not
found, and a new instance is attempted to be created. This causes a DB
uniqueness constraint to fail.