[#2962] Hide honeypot field on registration page
The style to hide the honeypot field on the registration page went missing very soon after the field was added. Quick look at the git commit history doesn't explain exactly what happened, looks like for some reason the added CSS was removed in a later commit that had an older parent than the one where the style was added.