[#2933] Fix codelist_choices
utils.codelist_choices() acquired a bug when translating the codelists was introduced.

The reason for this is that codelist_choices assumes the "code" column always is the first one. This is not certain when the codelists are translated.

Fix by always finding the position of the "code" and "name" columns of the codelist.