[#2951] Hack to work around a bug in DRF
unique_together constraint seems to make the fields that are a part of it
hidden, even when they are marked as read_only. This commit adds a work around,
to fix this bug. Using the patch from
https://github.com/encode/django-rest-framework/pull/4192/ didn't seem to work.
There are other changes, possibly, that cause this code to not work in the
version of DRF we are using.