[#2995] Recalculate the country when lat/lng change
To keep the save fast, previously country was recomputed only when there was no
country set previously. This leaves a user with no way to update the country, if
an incorrect lat/lng had been entered.

Also, given the way project editor's save works correctly, latitude alone is
saved first, and the longitude is updated on the location, in a succeeding step.
This causes an incorrect location to be set on the project first, and that never
gets updated.

Closes #2995