[#3012] Fix problem with switching back to results tab
When narrative summary accordion is fully collapsed, switching back to the
results tab would fail because of an assumption about the values in keys all
being lists.  Since, the collapse element for narrative summaries is an
accordion - only one of the panels can be open at a time, the element is not a
list, and is undefined when no panels are open.