[#3026] Postgres views of aggregated values
First cut of a Postgres view aggregating a value.

The view calculates the "actual value" of an indicator period, i.e. the sum of all value fields on the period's approved updates.

Also includes a management command script. This is a "fixed" version of the manangement command included in django-pgviews, since the original script only works with Django 1.9+.