[#3040] Add new IndicatorDimension framework to project editor
The dimensions are now added on the project level and are then added to indicators.

NOTE: when a new dimension has been created, the project editor needs to be reloaded for it to be present in the indicators.
(cherry picked from commit 200ca4d8180ab2a9962006423c8eed9dee4ea060)