[#3040] Refactor Results framework client side code
Fetch and use the IndicatorDimensionName and IndicatorDimensionValue models in stead of IndicatorDimension.

Add basic selectors for the new models.

Add selector getDimensionNameValueIds.

Refactor prepareUpdateForm() in App.componentWillReceiveProps() to use the new models.

Refactor how dimension and disaggregation data is assembled and sent down the component tree.
A compound datas tructure, dimensionsAndDisaggs, is created in App.render() and sent down the component tree to finally be used in the DisaggregatedInputs component.