[#3126] Handle text filter as a special case since it is free form
The text filter allows users to enter free form text, which could result in no
projects being found for the given text. Other fields only allow selecting from
a list of options, and for every combination that is shown to users and
selectable by them, at least one project exists.

When no projects are returned for a given search string, if the text search is
not handled separately, the options for all the other filters are empty, and
this causes the filters to get cleared automatically. This is very weird UX.