[#3122] First cut of versioned API
Add versions to the API. We've had `v1/` as part of the API path since its inception and it's now possible to change the version of an endpoint (to `v2/`, `v3/` etc).

Add `rest/v2/result` endpoint.

Initial case is the `result/` endpoint. v1 has a pair of fields that require one DB request per record, making it very slow for projects with many results.