[#3205] Fix subtle bug with user & enumerator permissions for None
The permission predicates for is_org_user and is_org_enumerator were returning
True for None objects without checking if the user has a corresponding
employment of that kind. This isn't a problem for the User group, because users
in the User group have the least permissions, so this bug didn't really manifest
anywhere. When the enumerator role is added, though, users in the Users group,
but not in the Enumerators group will get permissions to add indicator updates,
for example.