[#3232] New Project methods
Methods to support single indicator period project hierarchies.

Add Project.ancestor() that returns the "top" project of a project hierarchy

Add Project.uses_single_indicator_period() that returns True if the project is part of a single indicator period project hierarchy.

Add Projet.project_dates() that returns the start and end date of the project, the actuals if available otherwise the planned.