[#3217] Create user_projects_access API endpoint
The /rest/v1/user_projects_access endpoint is used to support the limiting of project access to a subset of all projects a user has access to through association user -> employments -> organisations -> projects.

The fields of the endpoint are "id", "all_projects" and "user_projects"
  "id": the id of the User
  "all_projects": a list of all projects the user has access to through association as per above.
  "user_projects": a list that is a subset of the all_projects list, representing the projects the user actually has access to. If user_projects us null, all projects are available. If user_projects is and empty array, [], no projects are accessible.

For each project in the "all_projects" list the project ID and title fields are included.

The "user_projects" field is a list of project IDs only.

Note that while the endpoint supports listing, it is primarly meant to be used to administer project access for individual users and you can only PUT lists of projects to individual users.

The "user_projects" field can be updated by PUTting a JSON object with the following format:

{
    "user_projects": {
        "projects": [<project_id_1>, <project_id_2>, <project_id_3>, ...]
    }
}

This list _replaces_ the preivous list of projects.