[#3278] Update singe_period_indicators management script
Change the behaviour slightly when more than one period is encountered.

If there's more than one period, check if only one if "modified". If so keep the modified period and delete the others.

A period is modified if one of the fields target_value, target_comment, actual_value or actual_comment has a value or the period has at least one update.