[#2742] Fix race condition when creating new org from project editor
The update location and upload logo calls are made asynchronously, and the
server has a race condition where it over-writes the logo data written, when
updating the location.

This commit makes update location call synchronous, and updates the logo, only
after the location call has finished. This fixes the bug, though it might
actually be better to fix the server to handle such writes, better.

Closes #2742