[#3327] project directory map: Allow zoom levels less than 2
On smaller screens, the zoom level is sometimes set to 1, to be able to show all
the location markers on the map. We had some custom logic to prevent such zoomed
out maps, which doesn't seem to be working well.

This commit removes this logic and lets the maps API figure out what the best
zoom level for the current screen would be.

Closes #3327