[#3355] Add UserProject.include_restricted()
UserProject.include_restricted() is a class method that finds all restricted users eligible for access to new projects, based on employment with the reporting organisation and that the organisation has the include_restricted flag set to True.

UserProject.include_restricted() is called in ProjectViewSet.create()

Add tests checking that new projects are accessible in the correct way for both restricted and unrestricted users.