[#3355] Change include_restricted flag to enable_restrictions
When `enable_restrictions` for an organisation is off:

1. the admins of that organisation cannot restrict any users where the
organisation is a reporting partner. The users of the organisation, though, may
be restricted by admins of other organisations who have admin rights on those
users.

2. any new projects with that organisation as the reporting partner, become
accessible to all users in the organisations, even if they have been restricted
by users of another organisation.