[#3355] Disallow disabling restrictions if there's any restricted user
Checking for other admins who are able to restrict a user, before allowing
disabling of restrictions is quite hard and will turn out to be quite slow,
possibly. For now, we are going to disable the turning off of restrictions,
unless all the users employed by that organisation, and it's content owned
organisations have no restrictions enabled.