[#3382] Fix broken Full view/Overview
The Full view/Overview toggle in the results framework stops working after saving an update.

This is because the new update doesn't have any _meta member.

This is OK since there are no collapse object children, but utils.toggleTree() needs to be aware that some objects don't have a _meta.