[#3355] Fix bug where list of accessible project changes
When toggling the restrictions off and then on again the in user access management page, all projects were set to be inaccessible.

This behaviour is at least surprising and a more expected behaviour would be that the restrictions should remain the same as before the toggling was made.

The problem stems from unrestrict_projects() assuming that is_restricted is always True, but when reverting to restricting the user this is not the case.

Fix by adding the current state of is_restricted as an optional argument when calling unrestrict_projects().

Add test that checks this behaviour.