[#3357] Unify password forms validation
There are three forms for password entry, when you register, when you set your password in the recovery workflow and when you change it in MyRSR.

When registering and recovering Django froms are used, when changing the password in MyRSR the API is used.

Add a PasswordValidationMixin class that is used to validate the passwords in both forms for password entry.

Also use the actual validation functions defined in akvo.rsr.forms in the UserPasswordSerializer