[#3355] Add UserProjectsSerializer.may_unrestrict field
UserProjectsSerializer.may_unrestrict indicates if a user may be unrestricted.
Unrestricting a user is prevented when a user has restricted projects the admin cannot control.

Also add UserProjectAccessSerializer.validate() to prevent unrestricting.
The validation is done here rather than in UserProjectsSerializer since we need both the user and the admin objects.

Add tests for the may_unrestrict field and to check that changing of is_restricted can not take place if it's not allowed.