[#3293] Query for content_owned_orgs only once per request
The user object is created only once per request, and we can "cache" the result
of the content_owned_organisations on the user object for the request. This
avoids doing multiple queries that are super long. Speeding up the single query
is something that can be looked into later, if required.