[#3368] Fix missing gmaps_api_key in inclusion_tags templates
The extra_context is added only to template rendering for views. When rendering
the templates for inclusion_tags, the extra variables defined using
`extra_context` don't get passed. This causes the map URLs in the projects map
widget to show a warning about missing Google Maps API key.

This commit fixed the problem, that was assumed to have been fixed with
453088143f0f55967af4ded762089aca72bb2abd.