[#3414] Fix circular content ownership bug
When two organisations set each other as content_owner using the direct
`ForeignKey` relationship, an incorrect query actually left out such
organisations from each others' list of content owned organisations.

This commit corrects the query to filter out the correct organisations, but also
fixes a problem with recursive calls which didn't happen previously because of
the buggy query.

The tests also include a tests for another possible bug in the list of content
owned organisations. The fix for this bug can come in a different fix.