[#2384] Uncomment the __getattr__ function on the QuerySetManager
To be able to chain filters on querysets, the __getattr__ has been implemented,
but due to some differences between older versions of Django and 1.11 the
__getattr__ causes some problems. Adding a call to repr(queryset) seems to fix
the issue.

It seems like some kind of caching issue in Django while the apps are being
setup, but I don't exactly know what the bug is. Also, I'm not sure if this
might slow things down when the queryset is actually a big queryset.