[#3410] Script creating shadow organisations for EUTF
Add mgnmt script eutf_shadow_organisations.py.
This script finds all partners of the EUTF hierarchy projects and creates new, "shadow", organisations to replace the originals as partners in the projects of the hierarchy.

The reason for the shadows is to enable EUTF to control all employments that can access the hierarchy projects.
For this purpose all shadow organisations are set to be content owned by EUTF.

Add Organisation.original field. This OneToOne filed references the original organisation a shadow was copied from.
The Organisation.original field is used when we need to use the original organisations data. This is used in the IATI export.

Add migration for the Organisation.original field

Remove the RestrictedUserProjectsByOrg model and its migration.

Change migration IDs.

Modify the signals for Employment so that emails aren't generated when dealing with employments created for shadow organisations.