[3422] Add endpoint to import indicator
Add Project.import_indicator() method. This method takes a parent indicator ID and tries to import the indicator to the project, creating a Result object if needed.

Add tests for Project.import_indicator()

Rename variables in Project.add_indicator()

Add API endpoint /rest/v1/project/<id>/import_indicator/<id>/
This enpoint requires a project ID and an ID of a parent indicator, i.e. and indicator of a parent project that is to be imported.