[#3457] Let enumerators view private projects' results framework
Enumerators had change permissions on indicator period data (updates), but did
not have view permissions on the indicator periods, indicators or results that
are needed to populate this framework.

This commit adds a view_ permission on these hierarchical objects for the
enumerator. Everyone who has a `change` permission is assumed to have `view`
permissions already.

This commit also changes the code used to filter out objects in a queryset for
logged in users who are not admins. Previously, the objects were filtered based
on whether the user has `change` permissions on an object. This has now been
changed to check for `view` permission if the request is a `GET` and otherwise
check for the `change` permission.

Closes #3457