[#3231] Get rid of obj_data argument to update_object
We pass orig_data to update_object and this can be used to compute the obj_data.
Eventually, we'd like to get rid of this `pre_process_data` and delegate this
work to Django. But, it may break things currently since we may be doing some
additional processing that Django would not do.