[#3231] Remove pre process from create_or_update_objects_from_data
The pre-processing of data is done at individual function level, since we are
passing multiple fields to each of these functions, and would like to get rid of
the top level loop over the keys/fields in the data.

Eventually, we could possibly even remove the pre processing and let Django do
the right thing.