[#3231] Get rid of multiple loops to ensure all data is processed
We process the data in the hierarchy order processing the objects highest in the
hierarchy before the ones lower (Project -> Result -> Indicator ...). This
ensures that all the data is processed only after it's parent has already been
created. This lets us to get rid of the multiple loops over the submitted data.