[#3533] Add code to filter out inaccessible projects from a queryset
When displaying a list of projects to a user, we need to filter out the projects
inacessible to them, from the list. We again use the hierarchy organisation as
the "owner" of a project and restrict access based on whether a user is employed
by the owner or it's collaborator org, or not.