[#3533] Fix organisation groups for projects taken from hierarchy
Organisations groups were figured out by looking at the common organisations
where user and admin are employed, and are also partners of a project. This is a
problem for projects which are a part of a hierarchy where admin and user are
employed, but the hierarchy org is not a partner.

This commit fixes the bug by checking for hierarchy organisations for each
project, but may be this can be made a little faster with some query tricks!