[#3501] Add flag indicating if a period can get a new update
Percentage indicators can take only one update per indicator period, because
adding them up is tricky! To prevent users from adding multiple updates, we
raise an error on the server side, when saving the update. But, this makes the
UX very poor, especially given the error message isn't very helpful.

Instead, this commit now adds a flag on the period to indicate if it supports
adding another update. The clients can now use this flag to show/hide the option
to add a new update.