[#3501] Add new action type to update selected fields of a model
Actions of the type `UPDATE_MODEL_FULFILLED` only update a complete model
object, and it is useful to have a way of updating some fields only without
having access to the data of the other fields. `UPDATE_MODEL_FIELD` type action
lets you update few/selected fields of a model.