[#3586] Use key part ids to sort fields along with level
When creating a bunch of new indicators/results, the ordering of the indicators
and results is random and not the same as in which they were entered. This
commit fixes the ordering by sorting the fields based on the IDs of the fields,
along with the level of the model in the Results framework hierarchy.

Closes #3586