[#3591] Don't fetch unnecessary models for public results tab view
Reports, Narrative reports and categories for a project needn't be fetched in
the public results view of a project. The API end-point for the results of a
project was changed and it broke the fetching of results for a user who is not
signed in. This caused the results page to not load. This commit removes the API
calls for these models in the public results view, and loads the page as soon as
the required results data has been fetched.