[#3612] Setup handlers on dynamic vocabulary dropdowns only once
The code that is setting up the change handlers, and creating the dynamic
dropdowns, and switching between them is called each time a new partial is
added. This causes an existing selection in a dropdown to be removed.

This commit fixes that by ensuring that the dynamic dropdown is not setup
multiple times, if it has already been setup.

Closes #3612