[#3620] Always make the IATI suffix field read only
Currently, only EUTF is using this feature and they want the IATI suffix field
to be readonly. They want the field to be readonly, even before a parent has
been added to the project to mark it as a part of the hierarchy. This commit
sets the field to be read-only always. We can find a better way of making this
work for EUTF, if another partner wants to use this feature.