Link to an older version of countries.json to get builds working
The data changed at the link we were pointing to, causing our builds to
fail. This commit sets the URL to point to the data that was being
downloaded when the code for this was added. Just slightly older data from
today was working too, but a few attempts at this, didn't succeed.