[#3759] Switch existing reporting partner to implementing partner
For EUTF project, we automatically set the reporting partner to EUTF when a
project becomes a part of the EUTF hierarchy. In some cases, this causes the
user who created the project to lose access to the project they have
created. To fix this, this commit changes the existing reporting
organisation to an implementing partner, when setting EUTF as the reproting
organisation.

Closes #3759