[#3763] Move nginx health check to port 8081
So that in the training env we can deploy ReportServer in the same pod.

This simplifies the training environment as we avoid creating another
deployment, service and tying together RSR and RS through it.

Changing ReportServer to listen to another port is a lot more cumbersome
than just changing nginx healthz check.