[#3777] Run IATI checks on related objects CRUD using API
IATI checks were previously run when the project editor save was run. The
new Project Editor uses a different code path, and the IATI checks are not
being run. This commit modifies the PublicProjectViewSet CRUD methods to run
project IATI checks for all objects that are related to a project.