[#3742] Add LogEntry for all CRUD operations from the API
This commit adds logging in the Django `LogEntry` table for all API calls to
viewsets that are subclassing the `PublicProjectViewSet`. If the API call
updates an object other than a `Project`, two log entries are created - one
for the actual object that changed, and another for the project to which
this object is related to.

The function to log these changes is a cleaner version of the `log_changes`
function that was being used in the old Project Editor. The entry format has
been changed from the previous logging function to match the Django logging
format more closely.