[#3742] Change ProjectSerializer.last_modified_at field
Project.last_modified_at is updated internally by Django, when a save is
called on the project. `Project.last_modified_by` is a property that uses
entries in the `LogEntry` table to get last modified time and user
information. The `ProjectSerializer` has been updated in this commit to use
this property on the Project for the `last_modified_at` field.

Closes #3742