[#3796] Remove validation set rules from the admin UI
The validation rules are now defined in the frontend and the partnership
team is used to editing the validations in the admin interface. This can
cause confusion, because the changed rules will not be applied in the new
PE. This commit removes the rules from the admin interface, and adds a note
about contacting the dev team to update validation sets.

Closes #3796