[#3820] Upgrade the dev container to the latest Python version
The prod container hasn't been upgraded to the newest one. This should
ideally be done in the same commit, to prevent developers from working in a
different environment from the one being used in production.

But, splitting the commits would allow us to work with the new container
locally, for a while, before we decide to make the move.