[#3822] Allow deleting projects that have imported results
When the project's relation to its parent is defined on the project that is
being deleted, the `RelatedProject` object is prevented from being deleted,
because we prevent changing project relations after importing results into a
project. So, this commit overrides the `Project.delete` method to delete all
the results on the project, before trying to delete it.

Closes #3822