[#3839] Support dropdown type for custom fields in the DB
This commit adds a `JSONField` on the `OrganisationCustomField` and
`ProjectCustomField` to store dropdown options for the dropdowm type custom
fields. The PE UI now needs to display the right kind of UI element based on
the type of the custom field.

It also adds a separate field to store the dropdown selections, instead of
using the value field.