[#3839] Add backend support to filter projects by custom fields
The project directory API now supports filtering projects by custom fields
that are present on the projects.

The filtering has been added as a separate function from the other filtering
because this filtering is happening on related fields, and there can be
multiple custom filters in the query. This was causing problems with using Q
objects and constructing the query before doing any queries.