[#3839] Show projects without images/titles on partner sites
We exclude projects without images on the RSR site, but for partner sites,
it may be confusing. This commit displays all the projects that need to
appear on a site, on partner sites. Partners are expected to fix data of
"ugly" projects from the project editor to fix their sites.