[#3887] Add separate model to keep track of default periods
Storing the Indicator as the default period was convenient for the user, but
it was also confusing and hard to work with. Instead, we store default start
and end dates for periods in a separate model. We could try to store it in
an existing indicator, but it just complicates things in various scenarios
when a period with specified start and end dates does not exist, etc.