[#3810] Check user permissions on project when creating related objs
Previously, we didn't check if the user has permissions on the project
before creating objects that were related to a particular project. This
actually meant that users from an unrelated organisation could create
objects on projects that they can't otherwise view.

We had special code to handle this correctly for project updates, but not
for other kinds of objects. This commit generalizes that for all kinds of
objects that are related to projects.