[#3810] Include projects of explicitly content owned organisations
We include the projects where explicitly content owned organisations are a
partner. The users of the content owner have permissions to access these
projects, and it makes sense to include them in the list returned by
my_projects for a user.

For implicitly content owned organisations, only the projects that cause the
partnership to make the organisation a content owner should be included, and
they would be included by default because of the partnership.