[#2208] Run a custom 2to3 fixer to rename __unicode__ to __str__
The fixer was obtained from this [blog
post](http://lucumr.pocoo.org/2010/2/11/porting-to-python-3-a-guide/) and
was run as described in [this](https://stackoverflow.com/a/24512708) Stack
Overflow answer.

Also, manually removed all references to `__unicode__` methods. The DRF
serializers still refer to `__str__` methods, but that could possibly be
changed?