[#3950] Move call to Project.new_project_created to the serializer
The response on project creation doesn't contain the correct information in
the `primary_organisation` field, because the `create` method uses "stale"
serialized data that was created from a fresh project. On project creation,
we do some book keeping, that happens after this response data has been
serialized. This commit moves the book-keeping code to be called before the
project is serialized, so that the response to this POST contains the
complete data for the project.

Closes #3950