[#3954] Extract function user_filtered project_ids
The function returns a set of project IDs that the user has access to, given
the user, organisations they are employed in, and the hierarchy under which
we are looking in. The hierarchy has the setting for whether or not the
project has restricted permissions.

The function has been extracted so that we can add caching around this code
path, so that we don't hit the DB repeatedly.