[#3954] Add caching to user_filtered_project_ids
The function caches data for a very short period of time, since clearing the
cache is a big pain. We only cache the data so that the same data is not
computed in the same request, over and over, for the same user. This ensures
that specific requests where this data is computed a lot of times are much
simpler to deal with.