[#3983] Add disaggregation values to the IATI exports
This commit uses the IndicatorPeriodDisaggregation values to add the
disaggregation values on quantitative indicators to the IATI export. The
IATI standard allows adding actual element inside a period to appear any
number of times, which means the actual values can be reported multiple
times inside a period, with the different disaggregations.

The period in the test in this commit for example, looks like this in the
IATI export.

```xml
<period>
  <period-start iso-date="2020-02-19"/>
  <period-end iso-date="2020-02-20"/>
  <target value="10">
    <comment>
      <narrative>Comment</narrative>
    </comment>
  </target>
  <actual value="10.00"/>
  <actual value="4.00">
    <dimension value="Male" name="Gender"/>
  </actual>
  <actual value="6.00">
    <dimension value="Female" name="Gender"/>
  </actual>
  <actual value="5.00">
    <dimension value="under 18" name="Age"/>
  </actual>
  <actual value="5.00">
    <dimension value="18 or above" name="Age"/>
  </actual>
</period>
```

The period has 5 actual sub-elements, the first one reports the "total"
actual value, without the disaggregations. The remaining 4 actual
sub-elements report the actual values for each of the different
disaggregations, along the two different dimensions of age and gender.

The IATI export also allows having multiple dimension sub-elements for an
actual value, and if we had the data for combinations of the
disaggregations, we could have reported them. For instance, data for
"Female" and "18 or above" could be reported with another actual sub-element
that has two dimension sub-elements. But, our data model does not allow for
adding data disaggregated over multiple dimensions, as of now.

References:

1. http://reference.iatistandard.org/203/activity-standard/iati-activities/iati-activity/result/indicator/period/actual/
2. http://reference.iatistandard.org/203/activity-standard/iati-activities/iati-activity/result/indicator/period/actual/dimension/