[#4013] Create default periods only once in child indicators
When a new indicator is created at the parent level, the corresponding child
indicators are automatically created for it. But, we have two locations
where the periods are created on the child indicators - one is a hook to
ensure that all the default periods are present on the child indicators,
another is the code that ensures that the imported result framework is in
sync with the parent framework. This causes some periods to be duplicated.

This commit fixes the creation of duplicate periods, by ignoring the
creation of default period copies on new indicators, if they already have a
parent indicator. The code that ensures that the results framework stays in
sync with the parent framework will create the required periods on the child
indicator.

Closes #4013