[#4043] Ensure only user accessible projects are listed in Up app
Use the `User.my_projects` method to return the correct list of project IDs
accessible to a user after the sign in using the app, and obtain the auth
token.

This list is only refreshed when the user logs in, and the FAC could
potentially cause issues for users who have the list of projects cached in
their apps. But, I'm not sure there's a way to resolve it from the server
side, without making any app code changes. May be invalidating the tokens
for users might work, but it could break things for non-Up app users
unexpectedly. Not sure if the app itself handles expired tokens correctly.

Closes #4043