[#4058] Default ACL to publicRead
We changed the storage buckets to Unified-permissions as Google Transfer Service does not copy the ACL of the source bucket to the target bucket, so that we did not need to set the correct ACLs using gsutil after the TEST/Training environment sync, as it takes more than 10 mins to do it.

But using a Unified-permissions means that the URLs of the Google Storage are signed urls (even if I think they do not need to be), which makes the urls unique, which means that the thumbnails need to be recreated every few minutes or hours (did not investigate when the urls expire).

The thumbnails are being created on any GET request means that the request, if the thumbnails are not there, takes more than 10 secs, making the whole of RSR slow.

Changing the Google Storage buckets to fine-grained permissions and setting the default ACL to publicRead so that the Google Storage urls do not need to be signed.

Also, I hope that the urls then are the same as the urls in the Persistent Disk, so that the Thumbnails do not need to be all regenerated.